+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-4548: Fixed stage manager not refreshing when enabling the stage manager feature flag while a stage is open
 - REMIX-4542: Fixed updating particle gizmo transform when moved using transform manipulators
 - Fixed a crash on selection of prims compatible with particle systems
+- Fixed logic to enable particle system creation on prims
 
 ### Removed