+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created 1.2.2 build
 
 ### Changed
+- Updated hdremix and omni_core_materials to ext-2ad1c2e-release-v1-2
 
 ### Fixed
 - Fixed a crash on selection of prims compatible with particle systems